requirements.txt
streamlit==1.36.0
plotly==5.24.1
pandas==2.2.3
numpy==1.26.4
shapely==2.1.0
scipy==1.14.1
scikit-learn==1.6.0
ezdxf==1.4.0
PyMuPDF==1.24.14
opencv-python-headless==4.10.0.84
pillow==10.4.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
python-dotenv==1.0.1
reportlab==4.2.2
matplotlib==3.8.4
networkx==3.3
streamlit
ezdxf
matplotlib
networkx
numpy
opencv-python-headless
pandas
pillow
plotly
psycopg2-binary
PyMuPDF
python-dotenv
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
streamlit
asyncio
ezdxf
networkx
numpy
opencv-python
pandas
plotly
psycopg2-binary
pymupdf
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
websockets
data-validator
streamlit
asyncio
ezdxf
networkx
numpy
opencv-python
pandas
plotly
psycopg2-binary
pymupdf
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
websockets
streamlit
asyncio
ezdxf
networkx
numpy
opencv-python
pandas
plotly
psycopg2-binary
pymupdf
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
websockets
streamlit
asyncio
ezdxf
networkx
numpy
opencv-python
pandas
plotly
psycopg2-binary
pymupdf
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
websockets
streamlit
asyncio
ezdxf
networkx
numpy
opencv-python
pandas
plotly
psycopg2-binary
pymupdf
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
websockets
streamlit
asyncio
ezdxf
networkx
numpy
opencv-python
pandas
plotly
psycopg2-binary
pymupdf
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
websockets
cairosvg
svgwrite
cairosvg
ezdxf
matplotlib
networkx
numpy
opencv-python-headless
pandas
pillow
plotly
psycopg2-binary
PyMuPDF
python-dotenv
redis
reportlab
scikit-learn
scipy
shapely
sqlalchemy
streamlit
svgwrite
websockets